internal/radix51: restructure according to golang.org/wiki/TargetSpecific
