+++ internal/radix51/fe_amd64.go
@@ -2,15 +2,12 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// +build amd64,!noasm
+// +build amd64,!purego
 
 package radix51
 
-// Mul sets v = x * y and returns v.
-func (v *FieldElement) Mul(x, y *FieldElement) *FieldElement {
-	feMul(v, x, y)
-	return v
-}
-
 //go:noescape
 func feMul(out, a, b *FieldElement)
+
+//go:noescape
+func feSquare(out, x *FieldElement)
