--- internal/radix51/mul.go
@@ -1,18 +1,0 @@
-// Copyright (c) 2017 George Tankersley. All rights reserved.
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-package radix51
-
-import "unsafe"
-
-// mul64x64 multiples two 64-bit numbers and adds them to two accumulators.
-// This function is written to ensure it inlines. I am so sorry.
-func mul64x64(lo, hi, a, b uint64) (ol uint64, oh uint64) {
-	t1 := (a>>32)*(b&0xFFFFFFFF) + ((a & 0xFFFFFFFF) * (b & 0xFFFFFFFF) >> 32)
-	t2 := (a&0xFFFFFFFF)*(b>>32) + (t1 & 0xFFFFFFFF)
-	ol = (a * b) + lo
-	cmp := ol < lo
-	oh = hi + (a>>32)*(b>>32) + t1>>32 + t2>>32 + uint64(*(*byte)(unsafe.Pointer(&cmp)))
-	return
-}
