+++ internal/radix51/mul_compat.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// +build !go1.12
+// +build !go1.13
 
 package radix51
